Use uint_least32_t instead of std::size_t as row and col indices
The type currently used for row and column indices is A::size_type (or std::size_t on essentially all allocators). However, that is an overshoot as in practice matrices with ~2*10^19 rows/columns would not even be possible to be stored. The next smaller candidate is uint_least32_t whose max value is about 4 billions. That's more than enough for all sequential cases and saves memory storage and bandwidth.