Rebind block allocator and construct allocators in place
The main purpose of this is to follow the standard library containers where the allocator may be of a different type that the primary type. In such case, the allocator_type must be the same as the template argument, and every instance where the allocator is needed has to be rebound.

Note that this includes the rebound of the BlockVector allocator being forwarded to std::vector. But this is just because current libc++ for Apple seems to not implement this yet.